requirements.txt
ollama~=0.5.4
pylint~=3.3.8
PyPDF2~=3.0.1
python-docx~=1.2.0
chromadb~=1.1.0